+++ CHANGELOG.md
@@ -29,3 +29,4 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Add outbound rules
 - Update swap slot scripts
 - Remove pre-commit hooks
+- Cloudflare DNS records add